+++ WMN_screenshooter.py
@@ -1,27 +1,24 @@
-#!/usr/bin/python
-
 """
-    Author : Micah Hoffman (@WebBreacher)
+    Initial author : Micah Hoffman (@WebBreacher)
     Additions by: Mike Eriksson (@swedishmike)
-    Description : Takes each username from the web_accounts_list.json file and performs the lookup to see if the entry is still valid and tries to take a screenshot of the valid ones. 
+    Description : Takes each username from the wmn-data.json file and performs the lookup to see if the entry is still valid and tries to take a screenshot of the valid ones that matches the username. 
 """
+
 import argparse
-import json
+import urllib3
 import os
 import signal
 import sys
+import json
 import re
-import errno
-from rich import print
 from queue import Queue
 from threading import Thread
+from rich import print
 from selenium import webdriver
-from selenium.common.exceptions import TimeoutException
-from time import sleep
+from selenium.webdriver.common.by import By
 from datetime import datetime
-
-import requests
-from requests.packages.urllib3.exceptions import InsecureRequestWarning
+from time import sleep
+import httpx
 
 
 # Parse command line input
@@ -38,7 +35,7 @@ parser.add_argument(
 parser.add_argument(
     "-c",
     "--config",
-    help="The full path to the web_accounts_list.json file",
+    help="The full path to the wmn-data.json file. If it's in the same path as this program - just wmn-data.json works.",
     required=True,
 )
 parser.add_argument(
@@ -55,18 +52,23 @@ parser.add_argument(
     type=int,
     default=50,
 )
+parser.add_argument(
+    "-x",
+    "--xxx",
+    help="Whether or not to query XXX sites. Default is set to No.",
+    action="store_true",
+)
+
 args = parser.parse_args()
 
-###################
-# Variables && Functions
-###################
 # Set HTTP Header info.
 headers = {
-    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/86.0.4240.75 Safari/537.36",
+    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36",
     "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
     "Accept-Language": "en-US,en;q=0.5",
     "Accept-Encoding": "gzip, deflate",
 }
+
 # Regular expression to remove http/https from sites to use in filenames
 remove_url = re.compile(r"https?://?")
 
@@ -86,40 +88,23 @@ def signal_handler(*_):
     sys.exit(130)
 
 
-def web_call(location):
-    try:
-        # Make web request for that URL, timeout in X secs and don't verify SSL/TLS certs
-        resp = requests.get(
-            location,
-            headers=headers,
-            timeout=args.timeout,
-            verify=False,
-            allow_redirects=False,
-        )
-    except requests.exceptions.Timeout:
-        return "[bold red]      ! ERROR: CONNECTION TIME OUT. Try increasing the timeout delay.[/bold red]"
-    except requests.exceptions.TooManyRedirects:
-        return "[bold red]      ! ERROR: TOO MANY REDIRECTS. Try changing the URL.[/bold red]"
-    except requests.exceptions.RequestException as e:
-        return "[bold red]      ! ERROR: CRITICAL ERROR." + e + "[/bold red]"
-    else:
-        return resp
-
-
 def read_in_the_json_file(filelocation):
     # Attempt to read in the JSON file
     try:
         with open(filelocation) as data_file:
             data = json.load(data_file)
     except FileNotFoundError:
-        print(f"[bold red] Could not find the JSON file - {filelocation} [/bold red]")
-        print("[bold red] Exiting....[/bold red]")
+        print(
+            f"[bold red] [!] Could not find the JSON file - {filelocation} [/bold red]"
+        )
+        print("[bold red] [!] Exiting....[/bold red]")
         sys.exit(1)
     except json.decoder.JSONDecodeError:
-        print("[bold red] The Json configuration file did not parse correctly.[/bold red]")
-        print("[bold red] Exiting....[/bold red]")
+        print(
+            "[bold red] [!] The Json configuration file did not parse correctly.[/bold red]"
+        )
+        print("[bold red] [!] Exiting....[/bold red]")
         sys.exit(1)
-
     return data
 
 
@@ -128,7 +113,7 @@ def validate_site(i, site_queue):
 
     while True:
         site = site_queue.get()
-        url = site["check_uri"].replace("{account}", args.username)
+        url = site["uri_check"].replace("{account}", args.username)
         r = web_call(url)
         if isinstance(r, str):
             # We got an error on the web call
@@ -136,18 +121,35 @@ def validate_site(i, site_queue):
 
         else:
             # Analyze the responses against what they should be
-            code_match = r.status_code == int(site["account_existence_code"])
-            string_match = r.text.find(site["account_existence_string"]) >= 0
+            code_match = r.status_code == int(site["e_code"])
+            string_match = r.text.find(site["e_string"]) >= 0
 
             if args.username:
                 if code_match and string_match:
-                    print(f"[bold green][+] Found user at {url}[/bold green]")
+                    # print(f"[bold green][+] Found user at {url}[/bold green]")
                     all_found_sites.append(url)
                     # continue
         site_queue.task_done()
 
 
-def queues_and_threads(data):
+def web_call(location):
+    try:
+        # Make web request for that URL, timeout in X secs and don't verify SSL/TLS certs
+        resp = httpx.get(
+            location,
+            headers=headers,
+            timeout=args.timeout,
+            verify=False,
+        )
+
+    except Exception as e:
+        return "[bold red] [!] ERROR " + location + " -> " + str(e) + "[/bold red]"
+
+    else:
+        return resp
+
+
+def queues_and_threads(sitelist):
     # Setting up the threads, ready to query URL's.
     for i in range(num_of_threads):
         worker = Thread(
@@ -157,16 +159,23 @@ def queues_and_threads(data):
                 site_queue,
             ),
         )
-        worker.setDaemon(True)
+        worker.daemon = True
         worker.start()
     # Validating the data in the json file so we only try sites that are valid
-    for site in data["sites"]:
+    for site in sitelist["sites"]:
         if not site["valid"]:
-            print(f"[bold cyan] *  Skipping {site['name']} - Marked as not valid.[/bold cyan]")
-            continue
-        if not site["known_accounts"][0]:
             print(
-                f"[bold cyan] *  Skipping {site['name']} - No valid user names to test.[/bold cyan]"
+                f"[bold cyan] [*]  Skipping {site['name']} - Marked as not valid.[/bold cyan]"
+            )
+            continue
+        if not site["known"][0]:
+            print(
+                f"[bold cyan] [*]  Skipping {site['name']} - No valid user names to test.[/bold cyan]"
+            )
+            continue
+        if not args.xxx and site["cat"] == "XXXPORNXXX":
+            print(
+                f"[bold cyan] [*]  Skipping {site['name']} - category is XXXPORNXXX and you're running the script without the -x/-xxx parameter.[/bold cyan]"
             )
             continue
         site_queue.put(site)
@@ -175,7 +184,7 @@ def queues_and_threads(data):
 
 def grab_screenshots(all_found_sites):
     print(
-        "[bold green]\nTrying to capture screenshot(s) from the identified site(s) now.[/bold green]"
+        "[bold green]\n [-] Trying to capture screenshot(s) from the identified site(s) now.[/bold green]"
     )
     options = webdriver.ChromeOptions()
     options.add_argument("headless")
@@ -189,7 +198,7 @@ def grab_screenshots(all_found_sites):
 
     try:
         print(
-            f"[bold green]The screenshots will be stored in [/bold green][bold cyan]{image_directory}[/bold cyan]"
+            f"[bold green] [-] The screenshots will be stored in [/bold green][bold cyan]{image_directory}[/bold cyan]"
         )
         os.makedirs(image_directory)
     except OSError as e:
@@ -197,7 +206,7 @@ def grab_screenshots(all_found_sites):
             raise  # This was not a "directory exist" error..
 
     for site in all_found_sites:
-        print(f"[bold green]Capturing: {site}[/bold green]")
+        print(f"[bold green] [-] Capturing: {site}[/bold green]")
         filename = (
             remove_url.sub("", site)
             .replace("/", "")
@@ -211,7 +220,7 @@ def grab_screenshots(all_found_sites):
             sleep(2)
             driver.get_screenshot_as_file(image_directory + "/" + filename)
         except TimeoutException as e:
-            print(f"[bold red]Timed out when trying to reach: {site}[/bold red]")
+            print(f"[bold red] [!] Timed out when trying to reach: {site}[/bold red]")
             continue
     driver.close()
 
@@ -220,20 +229,19 @@ def main():
     # Add this in case user presses CTRL-C
     signal.signal(signal.SIGINT, signal_handler)
 
+    # Read in the sitelist
+    sitelist = read_in_the_json_file(args.config)
+
     # Suppress HTTPS warnings
-    requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
-
-    # Read in the Json file with site definitions
-    data = read_in_the_json_file(args.config)
-
-    # Create the threads and queues to check all the sites
-    queues_and_threads(data)
+    urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
+    # Set up the threads and queues
+    queues_and_threads(sitelist)
 
     # Check if there's any sites to take a screenshot of - if there are, take it.
     if all_found_sites:
         grab_screenshots(all_found_sites)
     else:
-        print("[bold yellow]No sites found[/bold yellow]")
+        print("[bold yellow] [:(] No sites found[/bold yellow]")
 
 
 if __name__ == "__main__":
